Split progress repository tests
